# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATLAS \u2014 Market Intelligence Engine",
    "description": "Jhaveri Intelligence Platform \u2014 Market \u2192 Sector \u2192 Stock \u2192 Decision",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/system/roadmap": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Roadmap",
        "description": "Roadmap \u2014 parsed roadmap.yaml joined with state.db. 10s cached.",
        "operationId": "roadmap_api_v1_system_roadmap_get",
        "parameters": [
          {
            "name": "evaluate_slow",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Evaluate Slow"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChunkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepResponse"
            },
            "type": "array",
            "title": "Steps"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_shipped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Shipped At"
          },
          "last_ship_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ship Ok"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneResponse"
            },
            "type": "array",
            "title": "Milestones",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "attempts",
          "updated_at",
          "steps"
        ],
        "title": "ChunkResponse"
      },
...
      "DemoGateResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "walkthrough": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Walkthrough"
          }
        },
        "type": "object",
        "required": [
          "url",
          "walkthrough"
        ],
        "title": "DemoGateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MilestoneResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "MilestoneResponse",
        "description": "One R/A/G dot in the per-chunk process strip. `status` is one of\ngreen | amber | red | pending; `detail` is a short hover-tooltip\nexplaining where the value came from."
      },
...
      "RollupResponse": {
        "properties": {
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "pct": {
            "type": "integer",
            "title": "Pct"
          }
        },
        "type": "object",
        "required": [
          "done",
          "total",
          "pct"
        ],
        "title": "RollupResponse"
      },
...
      "StepResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "check": {
            "type": "string",
            "title": "Check"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "check",
          "detail"
        ],
        "title": "StepResponse"
      },
...
      "SystemRoadmapResponse": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionResponse"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "versions"
        ],
        "title": "SystemRoadmapResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rollup": {
            "$ref": "#/components/schemas/RollupResponse"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkResponse"
            },
            "type": "array",
            "title": "Chunks"
          },
          "demo_gate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DemoGateResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "goal",
          "status",
          "rollup",
          "chunks"
        ],
        "title": "VersionResponse"
      },
...
    }
  }
}